extract amp-specific methods from PreactBaseElement
