+++ src/preact/base-element.js
@@ -1,5 +1,4 @@
 import {devAssert} from '#core/assert';
-import {ActionTrust_Enum} from '#core/constants/action-constants';
 import {AmpEvents_Enum} from '#core/constants/amp-events';
 import {Loading_Enum} from '#core/constants/loading-instructions';
 import {ReadyState_Enum} from '#core/constants/ready-state';
@@ -12,11 +11,7 @@ import {
 } from '#core/context';
 import {Deferred} from '#core/data-structures/promise';
 import {createElementWithAttributes, dispatchCustomEvent} from '#core/dom';
-import {
-  Layout_Enum,
-  applyFillContent,
-  isLayoutSizeDefined,
-} from '#core/dom/layout';
+import {applyFillContent} from '#core/dom/layout';
 import {MediaQueryProps} from '#core/dom/media-query-props';
 import {childElementByAttr, childElementByTag} from '#core/dom/query';
 import {PauseHelper} from '#core/dom/video/pause-helper';
@@ -108,29 +103,6 @@ const HAS_PASSTHROUGH = (def) => !!(def.passthrough || def.passthroughNonEmpty);
  * @template API_TYPE
  */
 export class PreactBaseElement extends BaseElement {
-  /** @override @nocollapse */
-  static R1() {
-    return true;
-  }
-
-  /** @override @nocollapse */
-  static requiresShadowDom() {
-    // eslint-disable-next-line local/no-static-this
-    return this['usesShadowDom'];
-  }
-
-  /** @override @nocollapse */
-  static usesLoading() {
-    // eslint-disable-next-line local/no-static-this
-    return this['loadable'];
-  }
-
-  /** @override @nocollapse */
-  static prerenderAllowed() {
-    // eslint-disable-next-line local/no-static-this
-    return !this.usesLoading();
-  }
-
   /** @param {!Element} element */
   constructor(element) {
     super(element);
@@ -143,15 +115,6 @@ export class PreactBaseElement extends BaseElement {
       },
       'onPlayingState': (isPlaying) => {
         this.updateIsPlaying_(isPlaying);
-      },
-      'onLoading': () => {
-        this.handleOnLoading();
-      },
-      'onLoad': () => {
-        this.handleOnLoad();
-      },
-      'onError': () => {
-        this.handleOnError();
       },
     });
 
@@ -229,28 +192,6 @@ export class PreactBaseElement extends BaseElement {
    * @return {!JsonObject|undefined}
    */
   init() {}
-
-  /** @override */
-  isLayoutSupported(layout) {
-    const Ctor = this.constructor;
-    if (Ctor['layoutSizeDefined']) {
-      return (
-        isLayoutSizeDefined(layout) ||
-        // This allows a developer to specify the component's size using the
-        // user stylesheet without the help of AMP's static layout rules.
-        // Bento components use `ContainWrapper` with `contain:strict`, thus
-        // if a user stylesheet doesn't provide for the appropriate size, the
-        // element's size will be 0. The user stylesheet CSS can use
-        // fixed `width`/`height`, `aspect-ratio`, `flex`, `grid`, or any
-        // other CSS layouts coupled with `@media` queries and other CSS tools.
-        // Besides normal benefits of using plain CSS, an important feature of
-        // using this layout is that AMP does not add "sizer" elements thus
-        // keeping the user DOM clean.
-        layout == Layout_Enum.CONTAINER
-      );
-    }
-    return super.isLayoutSupported(layout);
-  }
 
   /** @override */
   buildCallback() {
@@ -372,22 +313,6 @@ export class PreactBaseElement extends BaseElement {
     }
   }
 
-  /** @override */
-  attemptChangeHeight(newHeight) {
-    return super.attemptChangeHeight(newHeight).catch((e) => {
-      // It's okay to disable this lint rule since we check that the restricted
-      // method exists.
-      // eslint-disable-next-line local/restrict-this-access
-      if (this.getOverflowElement && !this.getOverflowElement()) {
-        console./* OK */ warn(
-          '[overflow] element not found. Provide one to enable resizing to full contents.',
-          this.element
-        );
-      }
-      throw e;
-    });
-  }
-
   /**
    * @protected
    * @param {!JsonObject} props
@@ -403,24 +328,6 @@ export class PreactBaseElement extends BaseElement {
    */
   api() {
     return devAssert(this.currentRef_);
-  }
-
-  /**
-   * Register an action for AMP documents to execute an API handler.
-   *
-   * This has no effect on Bento documents, since they lack an Actions system.
-   * Instead, they should use `(await element.getApi()).action()`
-   * @param {string} alias
-   * @param {function(!API_TYPE, !../service/action-impl.ActionInvocation)} handler
-   * @param {../action-constants.ActionTrust_Enum} minTrust
-   * @protected
-   */
-  registerApiAction(alias, handler, minTrust = ActionTrust_Enum.DEFAULT) {
-    this.registerAction?.(
-      alias,
-      (invocation) => handler(this.api(), invocation),
-      minTrust
-    );
   }
 
   /**
@@ -511,43 +418,6 @@ export class PreactBaseElement extends BaseElement {
     if (this.resetLoading_) {
       this.resetLoading_ = false;
       this.mutateProps({'loading': Loading_Enum.AUTO});
-    }
-  }
-
-  /**
-   * Default handler for onLoad event
-   * Displays loader. Override to customize.
-   * @protected
-   */
-  handleOnLoad() {
-    this.toggleLoading?.(false);
-    this.toggleFallback?.(false);
-    this.togglePlaceholder?.(false);
-  }
-
-  /**
-   * Default handler for onLoading event
-   * Reveals loader. Override to customize.
-   * @protected
-   */
-  handleOnLoading() {
-    this.toggleLoading?.(true);
-  }
-
-  /**
-   * Default handler for onError event
-   * Displays Fallback / Placeholder. Override to customize.
-   * @protected
-   */
-  handleOnError() {
-    this.toggleLoading?.(false);
-    // If the content fails to load and there's a fallback element, display the fallback.
-    // Otherwise, continue displaying the placeholder.
-    if (this.getFallback?.()) {
-      this.toggleFallback?.(true);
-      this.togglePlaceholder?.(false);
-    } else {
-      this.togglePlaceholder?.(true);
     }
   }
